modules/nlp_control_plane/requirements.txt
# NLP Control Plane Module Dependencies
# Install: pip install -r requirements.txt

fastapi>=0.109.0
pydantic>=2.0.0
uvicorn>=0.27.0
pyyaml>=6.0
